+++ inject_func/inject_func_call.h
@@ -1,17 +1,15 @@
-#ifndef HW1_H
-#define HW1_H
+#ifndef INJECT_FUNC_CALL_H
+#define INJECT_FUNC_CALL_H
 
 #include "llvm/Pass.h" 
-#include "llvm/IR/Module.h"
 #include "llvm/IR/PassManager.h"
-#include "llvm/Support/raw_ostream.h"
 
 //------------------------------------------------------------------------------
 // Legacy PM interface
 //------------------------------------------------------------------------------
-struct Hello : public llvm::ModulePass {
+struct InjectFuncCall : public llvm::ModulePass {
   static char ID;
-  Hello() : llvm::ModulePass(ID) {}
+  InjectFuncCall() : llvm::ModulePass(ID) {}
   bool runOnModule(llvm::Module &M) override;
 };
