+++ inject_func/test.h
@@ -1,0 +1,6 @@
+#ifndef TEST_H
+#define TEST_H
+
+int baz(int a, int b);
+
+#endif
