+++ inject_func/test.ll
@@ -1,0 +1,96 @@
+; ModuleID = 'test.c'
+source_filename = "test.c"
+target datalayout = "e-m:o-p270:32:32-p271:32:32-p272:64:64-i64:64-f80:128-n8:16:32:64-S128"
+target triple = "x86_64-apple-macosx10.15.0"
+
+; Function Attrs: noinline nounwind ssp uwtable
+define i32 @foo(i32 %0) #0 {
+  %2 = alloca i32, align 4
+  store i32 %0, i32* %2, align 4
+  %3 = load i32, i32* %2, align 4
+  %4 = add nsw i32 %3, 1
+  ret i32 %4
+}
+
+; Function Attrs: noinline nounwind ssp uwtable
+define i32 @bar(i32 %0) #0 {
+  %2 = alloca i32, align 4
+  store i32 %0, i32* %2, align 4
+  %3 = load i32, i32* %2, align 4
+  %4 = add nsw i32 %3, 2
+  ret i32 %4
+}
+
+; Function Attrs: noinline nounwind ssp uwtable
+define i32 @main() #0 {
+  %1 = alloca i32, align 4
+  %2 = alloca i32, align 4
+  %3 = alloca i32, align 4
+  %4 = alloca [2 x i32], align 4
+  %5 = alloca [2 x i32], align 4
+  %6 = alloca i32, align 4
+  %7 = alloca i32, align 4
+  store i32 0, i32* %1, align 4
+  %8 = call i32 @foo(i32 1)
+  store i32 %8, i32* %2, align 4
+  %9 = call i32 @bar(i32 2)
+  store i32 %9, i32* %3, align 4
+  store i32 0, i32* %7, align 4
+  br label %10
+
+10:                                               ; preds = %28, %0
+  %11 = load i32, i32* %7, align 4
+  %12 = icmp slt i32 %11, 2
+  br i1 %12, label %13, label %31
+
+13:                                               ; preds = %10
+  %14 = load i32, i32* %7, align 4
+  %15 = load i32, i32* %7, align 4
+  %16 = call i32 @bar(i32 %15)
+  %17 = add nsw i32 %14, %16
+  %18 = load i32, i32* %7, align 4
+  %19 = sext i32 %18 to i64
+  %20 = getelementptr inbounds [2 x i32], [2 x i32]* %4, i64 0, i64 %19
+  store i32 %17, i32* %20, align 4
+  %21 = load i32, i32* %7, align 4
+  %22 = load i32, i32* %7, align 4
+  %23 = call i32 @foo(i32 %22)
+  %24 = add nsw i32 %21, %23
+  %25 = load i32, i32* %7, align 4
+  %26 = sext i32 %25 to i64
+  %27 = getelementptr inbounds [2 x i32], [2 x i32]* %5, i64 0, i64 %26
+  store i32 %24, i32* %27, align 4
+  br label %28
+
+28:                                               ; preds = %13
+  %29 = load i32, i32* %7, align 4
+  %30 = add nsw i32 %29, 1
+  store i32 %30, i32* %7, align 4
+  br label %10, !llvm.loop !5
+
+31:                                               ; preds = %10
+  %32 = getelementptr inbounds [2 x i32], [2 x i32]* %4, i64 0, i64 1
+  %33 = load i32, i32* %32, align 4
+  %34 = getelementptr inbounds [2 x i32], [2 x i32]* %5, i64 0, i64 1
+  %35 = load i32, i32* %34, align 4
+  %36 = call i32 @baz(i32 %33, i32 %35)
+  store i32 %36, i32* %6, align 4
+  %37 = load i32, i32* %2, align 4
+  ret i32 %37
+}
+
+declare i32 @baz(i32, i32) #1
+
+attributes #0 = { noinline nounwind ssp uwtable "frame-pointer"="all" "min-legal-vector-width"="0" "no-trapping-math"="true" "stack-protector-buffer-size"="8" "target-cpu"="penryn" "target-features"="+cx16,+cx8,+fxsr,+mmx,+sahf,+sse,+sse2,+sse3,+sse4.1,+ssse3,+x87" "tune-cpu"="generic" }
+attributes #1 = { "frame-pointer"="all" "no-trapping-math"="true" "stack-protector-buffer-size"="8" "target-cpu"="penryn" "target-features"="+cx16,+cx8,+fxsr,+mmx,+sahf,+sse,+sse2,+sse3,+sse4.1,+ssse3,+x87" "tune-cpu"="generic" }
+
+!llvm.module.flags = !{!0, !1, !2, !3}
+!llvm.ident = !{!4}
+
+!0 = !{i32 1, !"wchar_size", i32 4}
+!1 = !{i32 7, !"PIC Level", i32 2}
+!2 = !{i32 7, !"uwtable", i32 1}
+!3 = !{i32 7, !"frame-pointer", i32 2}
+!4 = !{!"Homebrew clang version 13.0.0"}
+!5 = distinct !{!5, !6}
+!6 = !{!"llvm.loop.mustprogress"}
