--- inputs/test.ll
@@ -1,46 +1,0 @@
-; ModuleID = 'inputs/test.c'
-source_filename = "inputs/test.c"
-target datalayout = "e-m:o-p270:32:32-p271:32:32-p272:64:64-i64:64-f80:128-n8:16:32:64-S128"
-target triple = "x86_64-apple-macosx10.15.0"
-
-; Function Attrs: noinline nounwind ssp uwtable
-define i32 @foo(i32 %0) #0 {
-  %2 = alloca i32, align 4
-  store i32 %0, i32* %2, align 4
-  %3 = load i32, i32* %2, align 4
-  %4 = add nsw i32 %3, 1
-  ret i32 %4
-}
-
-; Function Attrs: noinline nounwind ssp uwtable
-define i32 @bar(i32 %0) #0 {
-  %2 = alloca i32, align 4
-  store i32 %0, i32* %2, align 4
-  %3 = load i32, i32* %2, align 4
-  %4 = add nsw i32 %3, 2
-  ret i32 %4
-}
-
-; Function Attrs: noinline nounwind ssp uwtable
-define i32 @main() #0 {
-  %1 = alloca i32, align 4
-  %2 = alloca i32, align 4
-  store i32 0, i32* %1, align 4
-  %3 = call i32 @foo(i32 1)
-  %4 = call i32 @bar(i32 2)
-  %5 = add nsw i32 %3, %4
-  store i32 %5, i32* %2, align 4
-  %6 = load i32, i32* %2, align 4
-  ret i32 %6
-}
-
-attributes #0 = { noinline nounwind ssp uwtable "frame-pointer"="all" "min-legal-vector-width"="0" "no-trapping-math"="true" "stack-protector-buffer-size"="8" "target-cpu"="penryn" "target-features"="+cx16,+cx8,+fxsr,+mmx,+sahf,+sse,+sse2,+sse3,+sse4.1,+ssse3,+x87" "tune-cpu"="generic" }
-
-!llvm.module.flags = !{!0, !1, !2, !3}
-!llvm.ident = !{!4}
-
-!0 = !{i32 1, !"wchar_size", i32 4}
-!1 = !{i32 7, !"PIC Level", i32 2}
-!2 = !{i32 7, !"uwtable", i32 1}
-!3 = !{i32 7, !"frame-pointer", i32 2}
-!4 = !{!"Homebrew clang version 13.0.0"}
